feat: introduce configuration infra
